Splitting NASA9 from CEA
We're going to update and have the coefficients handled differently
by each.
